backend/requirements.txt
itchat-uos
requests
python-socketio